lend page display token supply

+++ src/state/utils.ts
@@ -4,8 +4,8 @@ import getDomain from '../utils/env';
 const web3 = new Web3(getDomain());
 
 export const formatBigNumber = (value) => {
-  return parseInt(web3.utils.fromWei(BigInt(value).toString())).toLocaleString(
-    'en-US'
+  return parseFloat(web3.utils.fromWei(BigInt(value).toString())).toLocaleString(
+    'en-US',
     // {
     //   minimumFractionDigits: 2,
     //   maximumFractionDigits: 2,
